# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Используется для запросов из браузера.",
        "title": "Анализ Принадлежности Текста к Жанру",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8082",
    "basePath": "/api/v1",
    "paths": {
...
        "/text-analysis-request": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "Для модератора - все заявки. Для создателя - только его заявки.",
                "tags": [
                    "Домен заявки на анализ текста"
                ],
                "summary": "Получить список заявок",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Фильтр по статусу ('черновик', 'сформирован', 'завершён', 'отклонён')",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список заявок",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/ds.AnalysisRequestDTO"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизован",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/text-analysis-request/icon": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "Возвращает ID текущей черновой заявки и количество жанров в ней.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Домен заявки на анализ текста"
                ],
                "summary": "Получить информацию о черновике",
                "responses": {
                    "200": {
                        "description": "Информация о черновике",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "analysis_request_id": {
                                    "type": "integer"
                                },
                                "genres_in_request_count": {
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/text-analysis-request/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "SessionCookie": []
                    }
                ],
                "tags": [
                    "Домен заявки на анализ текста"
                ],
                "summary": "Получить одну заявку по id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Заявка",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ds.AnalysisRequestDTO"
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизован",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/text-analysis-requests/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "Обновляет поле 'TextToAnalyse' черновой заявки.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Домен заявки на анализ текста"
                ],
                "summary": "Обновить черновик",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки (черновика)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новый текст для анализа",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.UpdateAnalysisRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное обновление",
                        "schema": {
                            "$ref": "#/definitions/ds.AnalysisRequestDTO"
                        }
                    },
                    "400": {
                        "description": "Неверный формат ID или данных",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Неавторизован",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера/Не является черновиком",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "Удаляет заявку (только черновик).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Домен заявки на анализ текста"
                ],
                "summary": "Удалить черновик заявки",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки (черновика)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Успешное удаление"
                    },
                    "400": {
                        "description": "Неверный формат ID",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Неавторизован",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера/Не является черновиком",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/text-analysis-requests/{id}/form": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "Переводит статус черновика на 'на модерации'.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Домен заявки на анализ текста"
                ],
                "summary": "Отправить заявку на модерацию",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки (черновика)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешная отправка",
                        "schema": {
                            "$ref": "#/definitions/ds.AnalysisRequestDTO"
                        }
                    },
                    "400": {
                        "description": "Неверный формат ID",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Неавторизован",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера/Не является черновиком",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/text-analysis-requests/{id}/process": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "SessionCookie": []
                    }
                ],
                "tags": [
                    "Домен заявки на анализ текста"
                ],
                "summary": "Завершить или отклонить заявку (Только для Модератора)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Действие ('complete' или 'reject')",
                        "name": "action",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновленная заявка",
                        "schema": {
                            "$ref": "#/definitions/ds.AnalysisRequestDTO"
                        }
                    },
                    "401": {
                        "description": "Неавторизован",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен (не модератор)",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Заявка не найдена",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "ds.AnalysisGenreDTO": {
            "type": "object",
            "properties": {
                "CommentToRequest": {
                    "type": "string"
                },
                "GenreID": {
                    "type": "integer"
                },
                "GenreImageURL": {
                    "type": "string"
                },
                "GenreName": {
                    "type": "string"
                },
                "ProbabilityPercent": {
                    "type": "integer"
                }
            }
        },
        "ds.AnalysisRequestDTO": {
            "type": "object",
            "properties": {
                "AnalysisRequestID": {
                    "type": "integer"
                },
                "AnalysisRequestStatus": {
                    "type": "string"
                },
                "CompletedAt": {
                    "type": "string"
                },
                "CreatedAt": {
                    "type": "string"
                },
                "CreatorLogin": {
                    "type": "string"
                },
                "FormedAt": {
                    "type": "string"
                },
                "Genres": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ds.AnalysisGenreDTO"
                    }
                },
                "ModeratorLogin": {
                    "type": "string"
                },
                "TextToAnalyse": {
                    "type": "string"
                }
            }
        },
...
        "ds.UpdateAnalysisRequestDTO": {
            "type": "object",
            "properties": {
                "TextToAnalyse": {
                    "type": "string"
                }
            }
        },
...
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Используется для запросов через Insomnia/Postman: \"Bearer \u003cJWT\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}